Added extra image buffer to attempt to remove tearing
Added extra image buffer to write the image to before displaying it to
the screen, in an attempt to remove the tearing that was occurring when
printing rectangles especially. This does not appear to visibly improve
the image, but I have branched this in order to make comparisons with
the original implementation.